genre shifter start (not tested)
Have the start to the code for the genre shifter, but need the GenreClassifier to use for this, as well as a trained AutoEncoder model. For things TODO, GenreClassifier for the loss function is set up in a way as if GenreClassifier does not split the clips up into 3 second chunks, and the GenreClassifier also needs to take in a desired genre for the loss, so a slightly different dataset generator is needed (with a desired output genre and input genre) needs to be added.